Limiting auras to uint8 specs. Spell Power shit. Heirloom shit.

+++ Trunk/src/hearthstone-world/SpellAuras.h
@@ -789,7 +789,7 @@ public:
 	Modifier * mod;
 	AreaAuraList targets;//this is only used for AA
 
-	uint32 m_auraSlot;
+	uint8 m_auraSlot;
 
 	uint32 m_castedItemId;
 	bool m_areaAura;		// Area aura stuff -> never passive.
